Refactor code to PEP-8 standard
