+++ src/calculator/operators/__init__.py
@@ -1,1 +1,3 @@
+'''This module is the interface for the operators.'''
+
 from .operator_mapper import operator_map
