+++ src/calculator/operators/operator_mapper.py
@@ -1,4 +1,6 @@
-from .arithmetic_operator import *
+'''This module registers the operators to their labels.'''
+
+from .arithmetic_operator import Add, Subtract, Multiply, Divide
 
 # Register the operator and their names here
